requirements.txt
# Django and REST framework
Django>=4.2.0
djangorestframework>=3.14.0

celery>=5.3.0
django-celery-beat>=2.5.0
redis>=4.5.0

python-dotenv>=1.0.0

django-filter>=23.2
django-cors-headers>=4.2.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
elasticsearch>=8.0.0
django-elasticsearch-dsl>=7.3.0